requirements.txt
fastapi
uvicorn
openai>=1.51.0
manim
python-dotenv
pydantic>=2